refactor: change button component files names
