+++ src/components/button/index.ts
@@ -1,1 +1,2 @@
-export * from './button';
+export { default as Button } from './button';
+export type { ButtonProps } from './button.types';
