--- src/components/index.ts
@@ -1,1 +1,0 @@
-export * from '@awesome-ui/components/button/button';
