+++ src/index.ts
@@ -1,2 +1,2 @@
 export { default as defaultTheme } from '@awesome-ui/theme/defaultTheme';
-export * from '@awesome-ui/components';
+export * from '@awesome-ui/components/button';
